Add options to customize the conversion
User of the plugin can now customize the classes that will be added to the
checkbox, containing list item and the list itself, optionally disabling
them altogether as well. The checkbox can also be enabled (disabled by
default), so it can be checked/unchecked by user. For deeper
customization, user can provide their own function to generate
the checkbox HTML.